requirements.txt
black                   # PEP8 formatter
black[jupyter]          # PEP8 formatter for Jupyter NB
click                   # CLI great interface
dill                    # Pickle substitute for big data sets
mlflow                  # ML tracker and more
notebook                # all Jupyter notebook needs
plotly                  # interactive plots (with JS)
pylint                  # great linter
pytest                  # unit tests
pytest-cov              # unit tests, cover a specific function
python-dotenv           # Environment variables
pyyaml                  # YAML support
requests                # HTTP requests
scikit-learn            # ML must-have
seaborn                 # plots
streamlit               # cool Python GUI
tarfile                 # tar archives
tensorflow              # DL, CPU only
tensorflow[and-cuda]    # DL with GPU
# torch                   # DL (pytorch)
# transformers            # Hugging Face (models, etc.)
